refactor: move `empty-options` test to `UglifyJsPlugin` test
